feat: HOC withAuthentication
